requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/

django_extensions
Django==1.8
django-markup-deprecated==0.0.3
Markdown==2.5.1
django-staticfiles==0.2.0
django-mailer==0.2a1.dev1
django-email-confirmation==0.2.dev4
django-timezones==0.2.dev1
django-ajax-validation==0.1.3
django-bleach==0.3.0
pytz==2010h
python-openid==2.2.5
django-uni-form==0.7.0
python-dateutil==1.4.1
docutils==0.6
Pygments==1.3.1
python-twitter==0.5
simplejson==2.0.9
vobject==0.8.1c
diff-match-patch==20090804
geopy==0.94
django-atomformat==0.1.1.dev1
django-oembed==0.1.1
django-sorting>=0.1
django-tagging-ext==0.3.5
django-photologue==2.3
django-markup==0.3
Pillow
python-ldap
django-auth-ldap==1.2.2
versiontools
django_compressor==1.4
fabric
django-bootstrap
psycopg2
django-bootstrap3

-e git+https://git@github.com/amarandon/pinax.git#egg=Pinax
-e git+https://git@github.com/amarandon/django-audiotracks.git#egg=django_audiotracks
